STY: Changed BLS to MLC and also number of bins optimization
The BLS method changed to MLC (MultipleLogisticClassification) since it is more generic and can be applied to any digital data.
A new number of bins optimization was added in order to help the number of bins used to find the histogram peaks. A recurrent increase of sqrt() function is used, until a limit of 3 is reached.
This may help for data that have small peaks in the histogram.