ENH: DiffusionEntropyMapping has variable to number of bins
Now it is availble to add an abritrary number of bins in the entropy mapping calculation. Thus, the entropy calculation can be evaluated by different number of bins criteria.
DebugMode flag was also added in order to remove unnecessary cout messages in the command line (depending on the user's choice)