ENH: DiffusionEntropyMapping has a new input data
An additional diffusion space mask can be added in the calculation process, which helps to restrict the diffusion entropy space to a meaniful description. 
Example: calculates the entropy only in the brain area (WM+GM+CSF) or even only in the WM area.
A binary 3D volume should be added which a single label value indicating the image area which the EV should be applied.